+++ src-tauri/Cargo.toml
@@ -19,6 +19,8 @@ tauri-build = { version = "2.5.2", features = [] }
 [dependencies]
 parking_lot = { workspace = true }
 anyhow = { workspace = true }
+tokio = { workspace = true }
+draft = { workspace = true }
 warp = { version = "0.4.2", features = ["server"] }
 open = "5.3.2"
 log = "0.4.28"
@@ -33,12 +35,6 @@ once_cell = { version = "1.21.3", features = ["parking_lot"] }
 port_scanner = "0.1.5"
 delay_timer = "0.11.6"
 percent-encoding = "2.3.2"
-tokio = { version = "1.48.0", features = [
-  "rt-multi-thread",
-  "macros",
-  "time",
-  "sync",
-] }
 serde = { version = "1.0.228", features = ["derive"] }
 reqwest = { version = "0.12.24", features = ["json", "cookies"] }
 regex = "1.12.2"
@@ -122,12 +118,19 @@ tauri-plugin-updater = "2.9.0"
 criterion = { workspace = true }
 
 [workspace.dependencies]
+draft = { path = "crates/draft" }
 parking_lot = { version = "0.12.5", features = [
   "hardware-lock-elision",
   "send_guard",
 ] }
 anyhow = "1.0.100"
 criterion = { version = "0.7.0", features = ["async_tokio"] }
+tokio = { version = "1.48.0", features = [
+  "rt-multi-thread",
+  "macros",
+  "time",
+  "sync",
+] }
 
 [workspace]
 members = ["crates/*"]
@@ -141,11 +144,6 @@ tauri-dev = []
 tokio-trace = ["console-subscriber"]
 clippy = ["tauri/test"]
 tracing = []
-
-[[bench]]
-name = "draft_benchmark"
-path = "benches/draft_benchmark.rs"
-harness = false
 
 [profile.release]
 panic = "abort"
