+++ src-tauri/crates/draft/Cargo.toml
@@ -14,3 +14,4 @@ anyhow = { workspace = true }
 
 [dev-dependencies]
 criterion = { workspace = true }
+tokio = { workspace = true }
