+++ src-tauri/src/cmd/verge.rs
@@ -1,5 +1,6 @@
 use super::CmdResult;
-use crate::{cmd::StringifyErr as _, config::IVerge, feat, utils::draft::SharedBox};
+use crate::{cmd::StringifyErr as _, config::IVerge, feat};
+use draft::SharedBox;
 
 /// 获取Verge配置
 #[tauri::command]
