+++ src-tauri/src/config/config.rs
@@ -5,10 +5,11 @@ use crate::{
     constants::{files, timing},
     core::{CoreManager, handle, service, tray, validate::CoreConfigValidator},
     enhance, logging, logging_error,
-    utils::{Draft, dirs, help, logging::Type},
+    utils::{dirs, help, logging::Type},
 };
 use anyhow::{Result, anyhow};
 use backoff::{Error as BackoffError, ExponentialBackoff};
+use draft::Draft;
 use smartstring::alias::String;
 use std::path::PathBuf;
 use tokio::sync::OnceCell;
