+++ src-tauri/src/utils/mod.rs
@@ -1,6 +1,5 @@
 pub mod autostart;
 pub mod dirs;
-pub mod draft;
 pub mod format;
 pub mod help;
 pub mod i18n;
@@ -15,5 +14,3 @@ pub mod server;
 pub mod singleton;
 pub mod tmpl;
 pub mod window_manager;
-
-pub use draft::Draft;
